% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RWrapper.R
\name{getSyntheticData}
\alias{getSyntheticData}
\title{Create Synthetic Dataset}
\usage{
getSyntheticData(df, seed = 0, niters = 1e+05, burnin = 30000,
  stride = 500, numModel = 10, verbose = FALSE, upperLimit = 9)
}
\arguments{
\item{df}{A dataframe}

\item{burnin}{An integer: Number of burnins}

\item{numModel}{An integer: Number of synthetic datasets required as output}

\item{verbose}{Set = TRUE for output debug information}

\item{upperLimit}{An integer: Defines "small count"}
}
\value{
Dataframe containing the synthetic datasets
}
\description{
\code{getSyntheticData} returns a number of synthetic datasets using code proposed in "Synthesizing Truncated Count Data for Confidentiality,"
and developed by Sam Hawala, Jerry Reiter and Quanli Wang.
}
\examples{
getSyntheticData(df)
getSyntheticData(df, upperLimit = 10)
}

