% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mtcdForSmallCount.R
\name{mtcdForSmallCount}
\alias{mtcdForSmallCount}
\title{Create Synthetic Dataset}
\usage{
mtcdForSmallCount(df, seed = 0, niters = 1e+05, burnin = 30000,
  stride = 500, numModel = 10, verbose = FALSE, upperLimit = 9)
}
\arguments{
\item{df}{A dataframe with three columns, the last column containing count to be synthesized}

\item{seed}{positive integer, random seed}

\item{niters}{positive integer, number of iterations for parameter generation}

\item{burnin}{a positive integer specifying the number of burnins}

\item{stride}{positive integer}

\item{numModel}{a positive integer specifying the number of synthetic datasets to be generated}

\item{verbose}{logical. Set = TRUE for output debug information}

\item{upperLimit}{a positive defining the "small count" threshold}
}
\value{
A dataframe containing the synthetic datasets and disclosure risk measures
}
\description{
Returns synthetic datasets from a Bayesian hierarchical model. This function wraps C++
code proposed in "Synthesizing Truncated Count Data for Confidentiality," and developed
by Sam Hawala, Jerry Reiter and Quanli Wang. References can be found in the package
description.
}
\section{Details}{
 [Add model detail here]
}
\examples{
getSyntheticData(df)
getSyntheticData(df, upperLimit = 10)
}

