% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticdata-package.R
\docType{package}
\name{syntheticdata-package}
\alias{syntheticdata}
\alias{syntheticdata-package}
\title{syntheticdata: Synthetic Clinical Data Generation}
\description{
Generates privacy-preserving synthetic clinical datasets and
validates their statistical fidelity. Designed for GDPR/HIPAA
compliant data sharing in multi-site clinical research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CuiweiG/syntheticdata}
  \item Report bugs at \url{https://github.com/CuiweiG/syntheticdata/issues}
}

}
\author{
\strong{Maintainer}: Cuiwei Gao \email{48gaocuiwei@gmail.com} [copyright holder]

}
\keyword{internal}
