% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SyntheticData.R
\docType{package}
\name{SyntheticData}
\alias{SyntheticData}
\alias{SyntheticData-package}
\title{Generate Synthetic Dataset}
\description{
This package provides an R wrapper function (\code{\link{getSyntheticData}}) for the statistical
model proposed in "Synthesizing Truncated Count Data for Confidentiality," which was originally written
in C++ and developed by Sam Hawala, Jerry Reiter and Quanli Wang.
}
\section{Background}{

To maintain confidentiality, national statistical agencies traditionally do not include
small counts in publicly released tabular data products. They typically delete these
small counts, or combine them with counts in adjacent table cells to preserve the totals
at higher levels of aggregation.  In some cases these suppression procedures result in
too much loss of information. To increase data utility and make more data publicly
available, Hawala et. al. created methods and software (in C++) to generate
synthetic values for the small counts from a Bayesian hierarchical model.  The software
generates synthetic data and computes several measures of disclosure risk, and was
applied by the Census Bureau in synthesizing small county-to-county migration counts.

This package aims to make the experience more "user-friendly" by providing an R-interface.
The original source C++ model is left unchanged aside from minor additions for R-compatibility. All inputs
and outputs from the model are R objects, as described in the documentation for \code{\link{getSyntheticData}}.
}

\section{Functions}{

This package contains one function, \code{\link{getSyntheticData}}, an R-wrapper that utilizes
the Rcpp package.
}

\section{Properties}{

\tabular{ll}{
Package: \tab SyntheticData\cr
License: \tab GPL (>=2)\cr
Version: \tab 0.1
}
}
\author{
John Doe \email{John@doe.org}
}
\references{
\url{http://sites.duke.edu/tcrn/research-projects/downloadable-software/}
}

